Use the default references same as the script system
